Update Android Emulator Path
Since Android SDK Tools 25.3.0 (March 2017) The Android Emulator was removed from the SDK Tools package and moved to a different SDK directory. 
For recent API versions (and depending on the OS architecture) running the old tools/emulator can cause errors because the different architecture versions of the emulator won't be found in the tools folder of the Android SDK.
Given that from August this year the minimum target version required for submitting apps to the Play store will be API 26, the path of the emulator should be updated.