Add explanation about the negative lookahead regex
There seems to be a few questions on Stackoverflow about Detox running the wrong test suite. This is actually due to the fact that the users are misinterpreting the regex that is displayed in the terminal. Adding it to the readme should help those newer users that are unaware of the regex.