deprecated `--file` and `--specs`
This change allows users to pass in args the same way they do with the test runner of their choice. It makes it much easier to manage this way and it doesn't take away any functionality that their framework might offer currently. For example mocha already has a `--file` option that does something else.