Wrap apkPath in quotes for Windows (spaces)
Windows users with spaces in their APK path face an issue where the path is broken by spaces. I faced this issue and tried escaping the space, but it turned out the solution was to remove the backslash escapes I added and use quotation marks, like "path". Apostrophes like 'path' don't work.

For reference,
https://android.stackexchange.com/questions/185951/script-to-pull-from-directory-containing-spaces

This fixed the problem for me and I'm now able to run `detox build`.

I considered other places in the code where the path could be wrapped in quotes, but this seemed like the simplest and only place necessary.