Implement view hierarchy dumping on demand and on error
When an invocation error occurs, an attempt to dump view hierarchy will be attempted, and on success, the resulting URL will be placed in “viewHierarchyURL”.

New invoke command “captureViewHierarchy” has been added, where “viewHierarchyURL” is expected to hold the URL to place the view hierarchy. The URL must have “path/name.viewhierarchy” syntax.