fix(iOS): fix visibility expectation error message.
Fix visibility error message to show a valid percentage description.

Before this fix, `15` percent value was printed as `1,500%`, instead as `15%`, since
`NSNumberFormatterPercentStyle` expects the value to be the percent divided by 100
(for example, in the `en_US` locale, the number 0.123 will be represented as 12%).

Since we are already using a percent value, we don't need a string formatter for the
representation.