[Genymotion SaaS] Update gmsaas start flag to fix wrong device timeout
When starting a device, --stop-when-inactive was added in the past when the global timeout didn't existed but only the front end inactivity timeout. When the global timeout was released, this flag still pointed to the front end inactivity timeout which is the wrong time out for devices started from gmsaas. Removing it solves the issue.